+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - Main window not visible after application startup when the system tray icon has been disabled (#458).
+- CPU Energy Performance Preference only shows `performance` as the available option when launching the application while the CPU governor is set to `performance` (#466).
 
 ### Added
 - Power profiles support for AMD RX 7600 (#447). Requires Linux 6.11 or later.